pyproject.toml
[tool.poetry]
name = "df1-library"
version = "0.1.0"
description = ""
authors = ["DanStaf <danstaf@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
